requirements.txt
fastapi 
uvicorn 
pydantic
langchain 
langgraph
requests
python-dotenv
